requirements.txt
google-genai
python-dotenv
requests
beautifulsoup4
playwright